Minor refactoring and json data storage
Data is now stored in /src/userdata/[file].json, as opposed to
/data/[file].json. Naturally, I also converted and moved our test data.
Also removed a couple now-unused folders, as well as adding a call to
load_data() in main.py before declaring the accessor and mutator
functions. Finally, I added comments around (but did not comment out)
some code that I believe is no longer necessary, but I didn't write, so
I have left it alone for the purpose of breaking as little as possible
with my changes.